Improve and refactor UI and UX (#26)
* unit tests

* remove favourites upon logout

* CSS: added scroll and delete butt for fav modal

* forgot to add scrollbar for fav list

* repeating username test

* Doesn't add repeated fav recipes on front end

* Save to favourite submits after clicking OK

* Change save to favourites alert

* Document public/js/favourites.js

* removed CheckRecords so automatically checks for userpass.json

* loginUtils.js unit tests

* Change object notation in favutils

* Add semicolons

* Recipe test current state

* changed alert messages

* editted alerts message

* Unit test fav utils

* Finish unit test fav utils

* Unit tests and delete function in favutils

* Unit tests and delete function in favutils

* Favourite deletes on the front end; issue: clicking delete still opens content

* Favourite deletes on the front end

* Merge branch 'master' of https://github.com/emilieez/2520-finalProject into dev-emi

# Conflicts:
#	public/css/style.css
#	tests/login.test.js

* Fix fridge

* Add delete recipe confirmation

* Make search results look more organized

+++ public/js/favourites.js
@@ -121,8 +121,20 @@ function addRecipeLabelBtn(recipe) {
     delRecipeBtn.className = "delFavBtn";
     delRecipeBtn.innerHTML = " delete";
     delRecipeBtn.onclick = function (ev) {
-        deleteRecipe(ev, recipe);
-        hideAllContents();
+        swal('Are you sure you want to remove this recipe from favourites?', {
+            buttons: {
+                cancel: "No",
+                catch: {
+                    text: "Yes",
+                    value: "remove"
+                }
+            }
+        }).then((value) => {
+            if (value === "remove") {
+                deleteRecipe(ev, recipe);
+                hideAllContents();
+            }
+        });
     };
 
     var hiddenDel = document.createElement('input');
@@ -164,14 +176,10 @@ function deleteRecipe(ev, recipe) {
             let delForm = $('#del-' + id[1]);
             delForm.on('submit', function (e) {
                 e.preventDefault();
-
                 $.ajax({
                     type: 'POST',
                     url: '/favourite/delete',
-                    data: delForm.serialize(),
-                    success: function () {
-                        swal('deleted');
-                    }
+                    data: delForm.serialize()
                 })
             });
             delForm.submit();
